rename shared js for filter match type dropdowns
