+++ src/controllers/index.js
@@ -2,7 +2,7 @@ import GlobalSearchController from './global_search_controller';
 import RemoteFormController from './remote_form_controller';
 import SearchController from './search_controller';
 import SearchDateFilterController from './search_date_filter_controller';
-import SearchLikeFilterController from './search_like_filter_controller';
+import SearchMatchTypeFilterController from './search_match_type_filter_controller';
 import ToastController from './toast_controller';
 
 export const radControllers = [
@@ -10,6 +10,6 @@ export const radControllers = [
   { id: 'remote-form', controller: RemoteFormController },
   { id: 'search', controller: SearchController },
   { id: 'search-date-filter', controller: SearchDateFilterController },
-  { id: 'search-like-filter', controller: SearchLikeFilterController },
+  { id: 'search-match-type-filter', controller: SearchMatchTypeFilterController },
   { id: 'toast', controller: ToastController }
 ];
